split flatlist component into separate file, make flatlist component a function
